Add SyncWithRedux adapter component injecting wallet state into Redux
Assuming that web3-react only provides access to its internal state
via hooks (which may or may not be true as of version 8, which uses
zustand internally), depending on the dApp's requirements and existing
codebase, the dApp's components may (or may not) benefit from
receiving web3 wallet connection state from Redux or another
full-blown state management solution, rather than relying on hooks for
composing business logic.

So as a proof of concept of a potential stop-gap workaround, to
decouple wallet state from app logic, add a SyncWithRedux adapter
component which takes serializable wallet state from web3-react and
injects it into the Redux store.

This approach was suggested in this ETHDubai 2022 talk:

  https://youtu.be/DcLsFThMHH4?t=775

and used as an intermediate step (which has since been retired) in
sushiswap-interface:

  https://github.com/sushiswap/sushiswap-interface/commit/6633a6b2358aac3738f8ef95532cf0ba60b00e4a#diff-6b5e5eaa5d17f26e8d758927d7764e33d5895e07e82bc08b5dbcbfb6f8a9644b